move search Pydantic objects to new file
